+++ contracts/libraries/Position.sol
@@ -30,7 +30,7 @@ library Position {
     function get(
         mapping(bytes32 => Info) storage positions,
         address owner,
-        uint112 id
+        uint104 id
     ) internal view returns (Info memory) {
         return positions[keccak256(abi.encodePacked(owner, id))];
     }
@@ -39,7 +39,7 @@ library Position {
     function set(
         mapping(bytes32 => Info) storage positions,
         address owner,
-        uint112 id,
+        uint104 id,
         Info memory position
     ) internal {
         positions[keccak256(abi.encodePacked(owner, id))] = position;
